# openapi.yaml
openapi: 3.0.1
info:
  title: API Gestion Mairie
  description: 'API REST pour la gestion des réservations de biens municipaux, des
    utilisateurs, des groupes et des pièces justificatives. Système de coordination
    pour la mairie.'
  contact:
    name: Équipe Mairie
    email: contact@mairie.fr
  version: 1.0.0
servers:
- url: http://localhost:8080
  description: Serveur de développement local
security:
- bearer-jwt: []
...
paths:
...
  /users/{id}/statut:
    patch:
      tags:
      - Utilisateurs
      summary: Valide le statut d'un utilisateur
      description: Change le statut d'un utilisateur (PENDING -> ACTIVE/REJECTED)
        - SECRETARY/ADMIN
      operationId: updateUserStatut
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              type: string
              enum:
              - ACTIF
              - INACTIF
              - SUSPENDU
              - PENDING
              - ACTIVE
              - REJECTED
        required: true
      responses:
        '200':
          description: Statut mis à jour avec succès
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDTO'
        '403':
          description: Accès refusé
        '404':
          description: Utilisateur non trouvé
      security:
      - bearer-jwt: []
...
components:
  schemas:
...
    UserDTO:
      type: object
      properties:
        id_user:
          type: string
          format: uuid
        nom:
          type: string
        prenom:
          type: string
        email:
          type: string
        telephone:
          type: string
        role:
          type: string
          enum:
          - ADMIN
          - USER
          - MODERATEUR
          - TUTORED
          - SECRETARY
        statut:
          type: string
          enum:
          - ACTIF
          - INACTIF
          - SUSPENDU
          - PENDING
          - ACTIVE
          - REJECTED
...
  securitySchemes:
    bearer-jwt:
      type: http
      description: 'Authentification JWT. Format: Bearer {token}'
      scheme: bearer
      bearerFormat: JWT